PTWP-70 feat: database_modeling

+++ src/main/java/com/fillkie/service/dto/TeamDetailDto.java
@@ -10,5 +10,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class TeamDetailDto {
     private String teamName;
-    private int headcount;
+    private Long headcount;
 }
